+++ NEWS.md
@@ -1,5 +1,13 @@
 # ggeffects 1.3.1
 
+## Changes
+
+* Better automatic handling of offset-terms, both for predictions and generating
+  plots with raw data. When the model formula contains an offset-term, and the
+  offset term is fixed at a specific value, the response variable is now
+  automatically transformed back to the original scale, and the offset-term is
+  added to the predicted values. A warning is printed when model contains transformed offset-terms that are not fixed, e.g. via the `condition` argument.
+
 ## Bug fixes
 
 * Fixed issue in `hypothesis_test()`, where the `by` argument did not work